fix use client with new pop
