+++ web/app/components/UsStocksHeader/Timer.tsx
@@ -1,15 +1,14 @@
-import { PopoverTrigger } from '@radix-ui/react-popover'
 import { Icon } from '@tendenz/icons'
 import { hoursUntilNextWeekdayHour } from '../../lib/hoursUntilNextWeekdayHour'
 import { FreeFloatButton } from '../FreeFloatButton'
-import { PopRoot, PopoverContentStyled } from '../Pop2'
+import { PopRoot, PopTrigger, PopoverContentStyled } from '../Pop2'
 
 export default function Timer() {
 	const timeTill = hoursUntilNextWeekdayHour(2)
 
 	return (
 		<PopRoot>
-			<PopoverTrigger asChild>
+			<PopTrigger asChild>
 				<FreeFloatButton className="group flex items-center gap-1.5 data-[state=open]:bg-slateA11 data-[state=open]:text-slate1 data-[state=open]:hover:bg-slateA11">
 					<Icon
 						name="radix-icons/clock"
@@ -17,7 +16,7 @@ export default function Timer() {
 					/>
 					<span className="text-sm font-light leading-none">{timeTill}h</span>
 				</FreeFloatButton>
-			</PopoverTrigger>
+			</PopTrigger>
 			<PopoverContentStyled color="slate">
 				<p>{timeTill} hours left until the next market update.</p>
 				<p className="mt-2">
