+++ web/app/components/UsStocksHeader/UsStocksHeader.tsx
@@ -1,23 +1,13 @@
-'use client'
-import { useSigmaYesterdayInfinite } from '../../lib/api/clientApi'
-import { npl } from '../../lib/naturalLanguageProcessing'
 import SettingsButton from './SettingsButton'
+import { UsStocksHeaderTitle } from './UsStocksHeaderTitle'
 
 import Timer from './Timer'
 
 export default function UsStocksHeader({}: {}) {
-	const { data } = useSigmaYesterdayInfinite()
-	const lastDate = data?.[0]?.[0]
-		? npl(data[0][0].last.date as string)
-		: 'yesterday'
-
 	return (
 		<div className="grid grid-cols-default">
 			<div className="col-start-2 mb-[1.5vh] flex items-end  justify-between gap-5">
-				<h2 className="text-4xl font-normal text-slate11">
-					{lastDate}
-					&apos;s anomalies
-				</h2>
+				<UsStocksHeaderTitle />
 
 				<div className="flex items-end gap-2">
 					<Timer />
